del ds

+++ src/views/Datasets.vue
@@ -39,7 +39,7 @@
         md4
         xl3
         v-for="dataset in datasets"
-        :key="dataset.id"
+        :key="dataset.data_set_name"
       >
         <v-card>
           <v-card-title>
@@ -62,10 +62,30 @@
               flat
               icon
               color="pink"
-              @click="deleteDataSet(`${dataset.id}`)"
+              @click="showDeleteDialog(`${dataset.data_set_name}`)"
             >
               <v-icon>delete</v-icon>
             </v-btn>
+            <v-dialog v-model="showDelete" max-width="200">
+              <v-card>
+                <v-card-text>
+                  Do you want to delete this data set?
+                </v-card-text>
+                <v-card-actions>
+                  <v-btn
+                    color="green darken-1"
+                    flat
+                    @click="showDelete = false"
+                  >
+                    No
+                  </v-btn>
+                  <v-btn color="green darken-1" flat @click="deleteDataSet">
+                    Yes
+                  </v-btn>
+                  <v-spacer></v-spacer>
+                </v-card-actions>
+              </v-card>
+            </v-dialog>
           </v-card-actions>
         </v-card>
       </v-flex>
@@ -80,7 +100,9 @@ export default {
   data() {
     return {
       fab: false,
-      dlgShow: false
+      dlgShow: false,
+      showDelete: false,
+      toDeletedId: -1
     };
   },
   mounted() {
@@ -101,17 +123,26 @@ export default {
     showDataSet(show) {
       this.dlgShow = show;
     },
-    async deleteDataSet(id) {
-      await fetch(
-        `${this.$store.state.backend_root_url}/datasets/api/v1/${id}/`,
-        {
-          method: "DELETE",
-          headers: new Headers({
-            Authorization: `token ${this.$store.state.token}`
-          })
-        }
-      ).catch(error => console.error(error));
-      this.$store.dispatch("getDataSets");
+    showDeleteDialog(id) {
+      this.toDeletedId = id;
+      this.showDelete = true;
+    },
+    async deleteDataSet() {
+      if (this.toDeletedId != -1) {
+        await fetch(
+          `${this.$store.state.backend_root_url}/datasets/api/v1/${
+            this.toDeletedId
+          }/`,
+          {
+            method: "DELETE",
+            headers: new Headers({
+              Authorization: `token ${this.$store.state.token}`
+            })
+          }
+        ).catch(error => console.error(error));
+        this.$store.dispatch("getDataSets");
+        this.showDelete = false;
+      }
     }
   }
 };
